+++ src/nwdiag/elements.py
@@ -50,6 +50,7 @@ class DiagramEdge(blockdiag.elements.DiagramEdge):
 class Network(blockdiag.elements.NodeGroup):
     basecolor = (185, 203, 228)
     linecolor = (0, 0, 0)
+    trunk_diameter: int
 
     @classmethod
     def set_default_linecolor(cls, color):
@@ -72,6 +73,8 @@ class Network(blockdiag.elements.NodeGroup):
         self.hidden = False
         self.colwidth = 1
         self.colheight = 1
+        self.trunk_diameter = None
+        print(f'Network creation : {self.id}.')
 
     @classmethod
     def create_anonymous(cls, nodes, attrs=None):
@@ -109,9 +112,32 @@ class Network(blockdiag.elements.NodeGroup):
 
         return label
 
-
 class Route(blockdiag.elements.DiagramEdge):
-    pass
+    basecolor = (255, 0, 0)
+
+    @classmethod
+    def clear(cls):
+        super(Route, cls).clear()
+        cls.basecolor = (255, 0, 0)
+
+    def __init__(self, node1, node2):
+        super(Route, self).__init__(node1, node2)
+        self.path = 'la'  # as Left-Above
+        self.pad = 2
+
+    def set_path(self, value):
+        value = value.lower()
+        if value not in ('la', 'lb', 'ra', 'rb'):
+            msg = "unknown route path: %s" % value
+            raise AttributeError(msg)
+        self.path = value
+
+    def set_pad(self, value):
+        value = int(value)
+        if value <= 0:
+            msg = "route pad should be positive: %s" % value
+            raise AttributeError(msg)
+        self.pad = value
 
 
 class Diagram(blockdiag.elements.Diagram):
@@ -119,6 +145,8 @@ class Diagram(blockdiag.elements.Diagram):
     _Network = Network
     _Route = Route
 
+    network_trunk_diameter: int
+   
     def set_default_linecolor(self, color):
         super(Diagram, self).set_default_linecolor(color)
         self._Network.set_default_linecolor(self.linecolor)
@@ -147,6 +175,7 @@ class Diagram(blockdiag.elements.Diagram):
         self.groups = []
         self.networks = []
         self.routes = []
+        self.network_trunk_diameter = None
 
     def set_external_connector(self, value):
         value = value.lower()